Catch up to around 51min, organizing
