+++ src/flyer.rs
@@ -2,95 +2,46 @@
 //! [video][1]
 //!
 //! [1]: https://youtu.be/gboGyccRVXI?si=2cvzvTBkKPhk_YZI
-#![allow(unused)]
-use anyhow::{bail, Context, Result};
-use serde::{Deserialize, Serialize};
-use std::io::{StdoutLock, Write};
+use anyhow::{Context, Result};
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use std::io::StdoutLock;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-struct Body {
+pub struct Body<Payload> {
     #[serde(rename = "msg_id")]
-    id: Option<usize>,
-    in_reply_to: Option<usize>,
+    pub id: Option<usize>,
+    pub in_reply_to: Option<usize>,
     #[serde(flatten)]
-    payload: Payload,
+    pub payload: Payload,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(tag = "type")]
-#[serde(rename_all = "snake_case")]
-enum Payload {
-    Echo {
-        echo: String,
-    },
-    EchoOk {
-        echo: String,
-    },
-    Init {
-        node_id: String,
-        node_ids: Vec<String>,
-    },
-    InitOk,
+pub struct Init {
+    pub node_id: String,
+    pub node_ids: Vec<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-struct Message {
-    src: String,
+pub struct Message<Payload> {
+    pub src: String,
     #[serde(rename = "dest")]
-    dst: String,
-    body: Body,
+    pub dst: String,
+    pub body: Body<Payload>,
 }
 
-struct EchoNode {
-    id: usize,
-}
-impl EchoNode {
-    pub fn step(&mut self, input: Message, output: &mut StdoutLock) -> Result<()> {
-        match input.body.payload {
-            Payload::Echo { echo } => {
-                let reply = Message {
-                    src: input.dst,
-                    dst: input.src,
-                    body: Body {
-                        id: Some(self.id),
-                        in_reply_to: input.body.id,
-                        payload: Payload::EchoOk { echo: echo },
-                    },
-                };
-                serde_json::to_writer(&mut *output, &reply)
-                    .context("serialize response to echo")?;
-                output.write_all(b"\n").context("trailing newline")?;
-                self.id += 1;
-            }
-            Payload::EchoOk { .. } => {}
-            Payload::Init { .. } => {
-                let reply = Message {
-                    src: input.dst,
-                    dst: input.src,
-                    body: Body {
-                        id: Some(self.id),
-                        in_reply_to: input.body.id,
-                        payload: Payload::InitOk,
-                    },
-                };
-                serde_json::to_writer(&mut *output, &reply)
-                    .context("serialize response to init")?;
-                output.write_all(b"\n").context("trailing newline")?;
-                self.id += 1;
-            }
-            Payload::InitOk => bail!("should never receive initok"),
-        }
-        Ok(())
-    }
+pub trait Node<Payload> {
+    fn step(&mut self, input: Message<Payload>, output: &mut StdoutLock) -> Result<()>;
 }
 
-pub fn run_flyer() -> Result<()> {
+pub fn run_flyer<S, Payload>(mut state: S) -> Result<()>
+where
+    S: Node<Payload>,
+    Payload: DeserializeOwned,
+{
     use std::io::{stdin, stdout};
     let stdin = stdin().lock();
     let mut stdout = stdout().lock();
-    let inputs = serde_json::Deserializer::from_reader(stdin).into_iter::<Message>();
-
-    let mut state = EchoNode { id: 0 };
+    let inputs = serde_json::Deserializer::from_reader(stdin).into_iter::<Message<Payload>>();
 
     for input in inputs {
         let input = input.context("maelstrom input failure")?;
@@ -98,6 +49,5 @@ pub fn run_flyer() -> Result<()> {
             .step(input, &mut stdout)
             .context("node step failure")?;
     }
-
     Ok(())
 }
